+++ src/index.js
@@ -24,7 +24,14 @@ import getCaretWord from './get-caret-word';
 import * as pressOn from './press-on';
 import typeText from './type-text';
 import isTextSelected from './is-text-selected';
-import deleteSelection from './delete-selection';
+import isDocumentActive from './is-document-active';
+import focusDocument from './focus-document';
+import * as moveCursorTo from './move-cursor-to';
+import * as remove from './remove';
+import * as select from './select';
+import {
+    addEventListener
+} from './mutation-observer';
 
 
 export {
@@ -47,5 +54,10 @@ export {
     pressOn,
     typeText,
     isTextSelected,
-    deleteSelection
+    isDocumentActive,
+    focusDocument,
+    remove,
+    moveCursorTo,
+    select,
+    addEventListener
 };