+++ types/index.d.ts
@@ -1,4 +1,4 @@
-// Type definitions for google-docs-utils 2.2
+// Type definitions for google-docs-utils 2.3
 // Project: https://github.com/Amaimersion/google-docs-utils/blob/master/README.md
 // Definitions by: Sergey Kuznetsov <https://github.com/Amaimersion>
 // Definitions: https://github.com/Amaimersion/google-docs-utils
@@ -9,6 +9,17 @@
 
 type QuerySelectorResult<T = HTMLElement> = T | null;
 type QuerySelectorAllResult<T = HTMLElement> = T[];
+
+type CtrlModificator = {
+    ctrlKey?: boolean;
+};
+type ShiftModificator = {
+    shiftKey?: boolean;
+};
+type CtrlShiftModificator = (
+    CtrlModificator &
+    ShiftModificator
+);
 
 interface GetSelectionResult {
     text: string;
@@ -33,6 +44,10 @@ interface GetCaretWordResult {
     text: string;
     indexStart: number;
     indexEnd: number;
+}
+
+interface GoogleDocsEvent {
+    type: string;
 }
 
 //#endregion
@@ -77,26 +92,74 @@ export function getCaret(): GetCaretResult | null;
 export function getCaretWord(): GetCaretWordResult | null;
 
 export const pressOn: {
-    Character: (char: string) => void,
+    Character: (char: string, modificator?: CtrlShiftModificator) => void,
     Space: () => void,
-    Delete: () => void,
-    Backspace: () => void,
+    Delete: (modificator?: CtrlModificator) => void,
+    Backspace: (modificator?: CtrlModificator) => void,
     Enter: () => void,
     Tab: () => void,
-    ArrowLeft: () => void,
-    ArrowRight: () => void,
-    ArrowUp: () => void,
-    ArrowDown: () => void,
+    ArrowLeft: (modificator?: CtrlShiftModificator) => void,
+    ArrowRight: (modificator?: CtrlShiftModificator) => void,
+    ArrowUp: (modificator?: CtrlShiftModificator) => void,
+    ArrowDown: (modificator?: CtrlShiftModificator) => void,
     Undo: () => void,
     Redo: () => void,
-    SelectAll: () => void,
-    PrintDialog: () => void
+    PrintDialog: () => void,
+    End: (modificator?: CtrlShiftModificator) => void,
+    Home: (modificator?: CtrlShiftModificator) => void,
+    Bold: () => void,
+    Italic: () => void,
+    Underline: () => void
 };
 
 export function typeText(text: string): void;
 
 export function isTextSelected(): boolean;
 
-export function deleteSelection(): boolean;
+export function isDocumentActive(): boolean;
+
+export function focusDocument(): boolean;
+
+export const remove: {
+    PrevWord: () => void,
+    NextWord: () => void,
+    Selection: () => boolean
+};
+
+export const moveCursorTo: {
+    PrevCharacter: () => void,
+    NextCharacter: () => void,
+    PrevLine: () => void,
+    NextLine: () => void,
+    PrevWord: () => void,
+    NextWord: () => void,
+    PrevParagraph: () => void,
+    NextParagraph: () => void,
+    LineStart: () => void,
+    LineEnd: () => void,
+    DocumentStart: () => void,
+    DocumentEnd: () => void
+};
+
+export const select: {
+    All: () => void,
+    PrevCharacter: () => void,
+    NextCharacter: () => void,
+    PrevWord: () => void,
+    NextWord: () => void,
+    PrevLine: () => void,
+    NextLine: () => void,
+    PrevParagraph: () => void,
+    NextParagraph: () => void,
+    TextBetweenCursorAndLineStart: () => void,
+    TextBetweenCursorAndLineEnd: () => void,
+    TextBetweenCursorAndDocumentStart: () => void,
+    TextBetweenCursorAndDocumentEnd: () => void
+};
+
+export function addEventListener(
+    type: string,
+    listener: (event: GoogleDocsEvent) => any
+): void;
 
 //#endregion